requirements.txt
arxiv>=2.1.0
google-generativeai>=0.3.0
requests>=2.28.0
